+++ .changeset/brave-bags-report.md
@@ -1,0 +1,5 @@
+---
+'@mysten/sui.js': minor
+---
+
+Remove deprecated types directory
