+++ sdk/graphql-transport/tsconfig.json
@@ -8,7 +8,7 @@
 		"rootDir": "src",
 		"paths": {
 			"@mysten/sui.js/client": ["../../typescrupt/src/client"],
-			"@mysten/sui.js/transactions": ["../../typescript/src/builder"]
+			"@mysten/sui.js/transactions": ["../../typescript/src/transactions"]
 		}
 	},
 	"references": [{ "path": "../wallet-standard" }, { "path": "../typescript" }]
