+++ sdk/graphql-transport/vitest.config.ts
@@ -17,7 +17,7 @@ export default defineConfig({
 		alias: {
 			'@mysten/bcs': new URL('../bcs/src', import.meta.url).toString(),
 			'@mysten/sui.js/transactions': new URL(
-				'../typescript/src/builder',
+				'../typescript/src/transactions',
 				import.meta.url,
 			).toString(),
 			'@mysten/sui.js': new URL('../typescript/src', import.meta.url).toString(),
