+++ sdk/typescript/src/transactions/hash.ts
